feat(onerm): update dashboard next week and current month
re #59
