+++ lib/application/exercise/exercise_bloc.freezed.dart
@@ -36,22 +36,6 @@ class _$ExerciseEventTearOff {
       ids: ids,
     );
   }
-
-  UpdateNextMonth updateNextMonth(
-      {@required int exerciseId, @required Month nextMonth}) {
-    return UpdateNextMonth(
-      exerciseId: exerciseId,
-      nextMonth: nextMonth,
-    );
-  }
-
-  UpdateNextWeek updateNextWeek(
-      {@required int exerciseId, @required WeekEnum nextWeek}) {
-    return UpdateNextWeek(
-      exerciseId: exerciseId,
-      nextWeek: nextWeek,
-    );
-  }
 }
 
 // ignore: unused_element
@@ -65,8 +49,6 @@ mixin _$ExerciseEvent {
     @required
         Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
     @required Result remove(List<int> ids),
-    @required Result updateNextMonth(int exerciseId, Month nextMonth),
-    @required Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
   });
   @optionalTypeArgs
   Result maybeWhen<Result extends Object>({
@@ -74,8 +56,6 @@ mixin _$ExerciseEvent {
     Result fetch(),
     Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
     Result remove(List<int> ids),
-    Result updateNextMonth(int exerciseId, Month nextMonth),
-    Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
     @required Result orElse(),
   });
   @optionalTypeArgs
@@ -84,8 +64,6 @@ mixin _$ExerciseEvent {
     @required Result fetch(Fetch value),
     @required Result selectionMode(SelectionMode value),
     @required Result remove(Remove value),
-    @required Result updateNextMonth(UpdateNextMonth value),
-    @required Result updateNextWeek(UpdateNextWeek value),
   });
   @optionalTypeArgs
   Result maybeMap<Result extends Object>({
@@ -93,8 +71,6 @@ mixin _$ExerciseEvent {
     Result fetch(Fetch value),
     Result selectionMode(SelectionMode value),
     Result remove(Remove value),
-    Result updateNextMonth(UpdateNextMonth value),
-    Result updateNextWeek(UpdateNextWeek value),
     @required Result orElse(),
   });
 }
@@ -192,15 +168,11 @@ class _$Add with DiagnosticableTreeMixin implements Add {
     @required
         Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
     @required Result remove(List<int> ids),
-    @required Result updateNextMonth(int exerciseId, Month nextMonth),
-    @required Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
   }) {
     assert(add != null);
     assert(fetch != null);
     assert(selectionMode != null);
     assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
     return add(exercise, oneRm);
   }
 
@@ -211,8 +183,6 @@ class _$Add with DiagnosticableTreeMixin implements Add {
     Result fetch(),
     Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
     Result remove(List<int> ids),
-    Result updateNextMonth(int exerciseId, Month nextMonth),
-    Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
     @required Result orElse(),
   }) {
     assert(orElse != null);
@@ -229,15 +199,11 @@ class _$Add with DiagnosticableTreeMixin implements Add {
     @required Result fetch(Fetch value),
     @required Result selectionMode(SelectionMode value),
     @required Result remove(Remove value),
-    @required Result updateNextMonth(UpdateNextMonth value),
-    @required Result updateNextWeek(UpdateNextWeek value),
   }) {
     assert(add != null);
     assert(fetch != null);
     assert(selectionMode != null);
     assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
     return add(this);
   }
 
@@ -248,8 +214,6 @@ class _$Add with DiagnosticableTreeMixin implements Add {
     Result fetch(Fetch value),
     Result selectionMode(SelectionMode value),
     Result remove(Remove value),
-    Result updateNextMonth(UpdateNextMonth value),
-    Result updateNextWeek(UpdateNextWeek value),
     @required Result orElse(),
   }) {
     assert(orElse != null);
@@ -313,15 +277,11 @@ class _$Fetch with DiagnosticableTreeMixin implements Fetch {
     @required
         Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
     @required Result remove(List<int> ids),
-    @required Result updateNextMonth(int exerciseId, Month nextMonth),
-    @required Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
   }) {
     assert(add != null);
     assert(fetch != null);
     assert(selectionMode != null);
     assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
     return fetch();
   }
 
@@ -332,8 +292,6 @@ class _$Fetch with DiagnosticableTreeMixin implements Fetch {
     Result fetch(),
     Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
     Result remove(List<int> ids),
-    Result updateNextMonth(int exerciseId, Month nextMonth),
-    Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
     @required Result orElse(),
   }) {
     assert(orElse != null);
@@ -350,15 +308,11 @@ class _$Fetch with DiagnosticableTreeMixin implements Fetch {
     @required Result fetch(Fetch value),
     @required Result selectionMode(SelectionMode value),
     @required Result remove(Remove value),
-    @required Result updateNextMonth(UpdateNextMonth value),
-    @required Result updateNextWeek(UpdateNextWeek value),
   }) {
     assert(add != null);
     assert(fetch != null);
     assert(selectionMode != null);
     assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
     return fetch(this);
   }
 
@@ -369,8 +323,6 @@ class _$Fetch with DiagnosticableTreeMixin implements Fetch {
     Result fetch(Fetch value),
     Result selectionMode(SelectionMode value),
     Result remove(Remove value),
-    Result updateNextMonth(UpdateNextMonth value),
-    Result updateNextWeek(UpdateNextWeek value),
     @required Result orElse(),
   }) {
     assert(orElse != null);
@@ -473,15 +425,11 @@ class _$SelectionMode with DiagnosticableTreeMixin implements SelectionMode {
     @required
         Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
     @required Result remove(List<int> ids),
-    @required Result updateNextMonth(int exerciseId, Month nextMonth),
-    @required Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
   }) {
     assert(add != null);
     assert(fetch != null);
     assert(selectionMode != null);
     assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
     return selectionMode(isInSelectionMode, selectedIds);
   }
 
@@ -492,8 +440,6 @@ class _$SelectionMode with DiagnosticableTreeMixin implements SelectionMode {
     Result fetch(),
     Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
     Result remove(List<int> ids),
-    Result updateNextMonth(int exerciseId, Month nextMonth),
-    Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
     @required Result orElse(),
   }) {
     assert(orElse != null);
@@ -510,15 +456,11 @@ class _$SelectionMode with DiagnosticableTreeMixin implements SelectionMode {
     @required Result fetch(Fetch value),
     @required Result selectionMode(SelectionMode value),
     @required Result remove(Remove value),
-    @required Result updateNextMonth(UpdateNextMonth value),
-    @required Result updateNextWeek(UpdateNextWeek value),
   }) {
     assert(add != null);
     assert(fetch != null);
     assert(selectionMode != null);
     assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
     return selectionMode(this);
   }
 
@@ -529,8 +471,6 @@ class _$SelectionMode with DiagnosticableTreeMixin implements SelectionMode {
     Result fetch(Fetch value),
     Result selectionMode(SelectionMode value),
     Result remove(Remove value),
-    Result updateNextMonth(UpdateNextMonth value),
-    Result updateNextWeek(UpdateNextWeek value),
     @required Result orElse(),
   }) {
     assert(orElse != null);
@@ -618,15 +558,11 @@ class _$Remove with DiagnosticableTreeMixin implements Remove {
     @required
         Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
     @required Result remove(List<int> ids),
-    @required Result updateNextMonth(int exerciseId, Month nextMonth),
-    @required Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
   }) {
     assert(add != null);
     assert(fetch != null);
     assert(selectionMode != null);
     assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
     return remove(ids);
   }
 
@@ -637,8 +573,6 @@ class _$Remove with DiagnosticableTreeMixin implements Remove {
     Result fetch(),
     Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
     Result remove(List<int> ids),
-    Result updateNextMonth(int exerciseId, Month nextMonth),
-    Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
     @required Result orElse(),
   }) {
     assert(orElse != null);
@@ -655,15 +589,11 @@ class _$Remove with DiagnosticableTreeMixin implements Remove {
     @required Result fetch(Fetch value),
     @required Result selectionMode(SelectionMode value),
     @required Result remove(Remove value),
-    @required Result updateNextMonth(UpdateNextMonth value),
-    @required Result updateNextWeek(UpdateNextWeek value),
   }) {
     assert(add != null);
     assert(fetch != null);
     assert(selectionMode != null);
     assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
     return remove(this);
   }
 
@@ -674,8 +604,6 @@ class _$Remove with DiagnosticableTreeMixin implements Remove {
     Result fetch(Fetch value),
     Result selectionMode(SelectionMode value),
     Result remove(Remove value),
-    Result updateNextMonth(UpdateNextMonth value),
-    Result updateNextWeek(UpdateNextWeek value),
     @required Result orElse(),
   }) {
     assert(orElse != null);
@@ -691,342 +619,6 @@ abstract class Remove implements ExerciseEvent {
 
   List<int> get ids;
   $RemoveCopyWith<Remove> get copyWith;
-}
-
-abstract class $UpdateNextMonthCopyWith<$Res> {
-  factory $UpdateNextMonthCopyWith(
-          UpdateNextMonth value, $Res Function(UpdateNextMonth) then) =
-      _$UpdateNextMonthCopyWithImpl<$Res>;
-  $Res call({int exerciseId, Month nextMonth});
-}
-
-class _$UpdateNextMonthCopyWithImpl<$Res>
-    extends _$ExerciseEventCopyWithImpl<$Res>
-    implements $UpdateNextMonthCopyWith<$Res> {
-  _$UpdateNextMonthCopyWithImpl(
-      UpdateNextMonth _value, $Res Function(UpdateNextMonth) _then)
-      : super(_value, (v) => _then(v as UpdateNextMonth));
-
-  @override
-  UpdateNextMonth get _value => super._value as UpdateNextMonth;
-
-  @override
-  $Res call({
-    Object exerciseId = freezed,
-    Object nextMonth = freezed,
-  }) {
-    return _then(UpdateNextMonth(
-      exerciseId: exerciseId == freezed ? _value.exerciseId : exerciseId as int,
-      nextMonth: nextMonth == freezed ? _value.nextMonth : nextMonth as Month,
-    ));
-  }
-}
-
-class _$UpdateNextMonth
-    with DiagnosticableTreeMixin
-    implements UpdateNextMonth {
-  const _$UpdateNextMonth({@required this.exerciseId, @required this.nextMonth})
-      : assert(exerciseId != null),
-        assert(nextMonth != null);
-
-  @override
-  final int exerciseId;
-  @override
-  final Month nextMonth;
-
-  @override
-  String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
-    return 'ExerciseEvent.updateNextMonth(exerciseId: $exerciseId, nextMonth: $nextMonth)';
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties
-      ..add(DiagnosticsProperty('type', 'ExerciseEvent.updateNextMonth'))
-      ..add(DiagnosticsProperty('exerciseId', exerciseId))
-      ..add(DiagnosticsProperty('nextMonth', nextMonth));
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other is UpdateNextMonth &&
-            (identical(other.exerciseId, exerciseId) ||
-                const DeepCollectionEquality()
-                    .equals(other.exerciseId, exerciseId)) &&
-            (identical(other.nextMonth, nextMonth) ||
-                const DeepCollectionEquality()
-                    .equals(other.nextMonth, nextMonth)));
-  }
-
-  @override
-  int get hashCode =>
-      runtimeType.hashCode ^
-      const DeepCollectionEquality().hash(exerciseId) ^
-      const DeepCollectionEquality().hash(nextMonth);
-
-  @override
-  $UpdateNextMonthCopyWith<UpdateNextMonth> get copyWith =>
-      _$UpdateNextMonthCopyWithImpl<UpdateNextMonth>(this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result add(Exercise exercise, OneRm oneRm),
-    @required Result fetch(),
-    @required
-        Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
-    @required Result remove(List<int> ids),
-    @required Result updateNextMonth(int exerciseId, Month nextMonth),
-    @required Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
-  }) {
-    assert(add != null);
-    assert(fetch != null);
-    assert(selectionMode != null);
-    assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
-    return updateNextMonth(exerciseId, nextMonth);
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result add(Exercise exercise, OneRm oneRm),
-    Result fetch(),
-    Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
-    Result remove(List<int> ids),
-    Result updateNextMonth(int exerciseId, Month nextMonth),
-    Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (updateNextMonth != null) {
-      return updateNextMonth(exerciseId, nextMonth);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result add(Add value),
-    @required Result fetch(Fetch value),
-    @required Result selectionMode(SelectionMode value),
-    @required Result remove(Remove value),
-    @required Result updateNextMonth(UpdateNextMonth value),
-    @required Result updateNextWeek(UpdateNextWeek value),
-  }) {
-    assert(add != null);
-    assert(fetch != null);
-    assert(selectionMode != null);
-    assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
-    return updateNextMonth(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result add(Add value),
-    Result fetch(Fetch value),
-    Result selectionMode(SelectionMode value),
-    Result remove(Remove value),
-    Result updateNextMonth(UpdateNextMonth value),
-    Result updateNextWeek(UpdateNextWeek value),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (updateNextMonth != null) {
-      return updateNextMonth(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class UpdateNextMonth implements ExerciseEvent {
-  const factory UpdateNextMonth(
-      {@required int exerciseId,
-      @required Month nextMonth}) = _$UpdateNextMonth;
-
-  int get exerciseId;
-  Month get nextMonth;
-  $UpdateNextMonthCopyWith<UpdateNextMonth> get copyWith;
-}
-
-abstract class $UpdateNextWeekCopyWith<$Res> {
-  factory $UpdateNextWeekCopyWith(
-          UpdateNextWeek value, $Res Function(UpdateNextWeek) then) =
-      _$UpdateNextWeekCopyWithImpl<$Res>;
-  $Res call({int exerciseId, WeekEnum nextWeek});
-
-  $WeekEnumCopyWith<$Res> get nextWeek;
-}
-
-class _$UpdateNextWeekCopyWithImpl<$Res>
-    extends _$ExerciseEventCopyWithImpl<$Res>
-    implements $UpdateNextWeekCopyWith<$Res> {
-  _$UpdateNextWeekCopyWithImpl(
-      UpdateNextWeek _value, $Res Function(UpdateNextWeek) _then)
-      : super(_value, (v) => _then(v as UpdateNextWeek));
-
-  @override
-  UpdateNextWeek get _value => super._value as UpdateNextWeek;
-
-  @override
-  $Res call({
-    Object exerciseId = freezed,
-    Object nextWeek = freezed,
-  }) {
-    return _then(UpdateNextWeek(
-      exerciseId: exerciseId == freezed ? _value.exerciseId : exerciseId as int,
-      nextWeek: nextWeek == freezed ? _value.nextWeek : nextWeek as WeekEnum,
-    ));
-  }
-
-  @override
-  $WeekEnumCopyWith<$Res> get nextWeek {
-    if (_value.nextWeek == null) {
-      return null;
-    }
-    return $WeekEnumCopyWith<$Res>(_value.nextWeek, (value) {
-      return _then(_value.copyWith(nextWeek: value));
-    });
-  }
-}
-
-class _$UpdateNextWeek with DiagnosticableTreeMixin implements UpdateNextWeek {
-  const _$UpdateNextWeek({@required this.exerciseId, @required this.nextWeek})
-      : assert(exerciseId != null),
-        assert(nextWeek != null);
-
-  @override
-  final int exerciseId;
-  @override
-  final WeekEnum nextWeek;
-
-  @override
-  String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
-    return 'ExerciseEvent.updateNextWeek(exerciseId: $exerciseId, nextWeek: $nextWeek)';
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties
-      ..add(DiagnosticsProperty('type', 'ExerciseEvent.updateNextWeek'))
-      ..add(DiagnosticsProperty('exerciseId', exerciseId))
-      ..add(DiagnosticsProperty('nextWeek', nextWeek));
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other is UpdateNextWeek &&
-            (identical(other.exerciseId, exerciseId) ||
-                const DeepCollectionEquality()
-                    .equals(other.exerciseId, exerciseId)) &&
-            (identical(other.nextWeek, nextWeek) ||
-                const DeepCollectionEquality()
-                    .equals(other.nextWeek, nextWeek)));
-  }
-
-  @override
-  int get hashCode =>
-      runtimeType.hashCode ^
-      const DeepCollectionEquality().hash(exerciseId) ^
-      const DeepCollectionEquality().hash(nextWeek);
-
-  @override
-  $UpdateNextWeekCopyWith<UpdateNextWeek> get copyWith =>
-      _$UpdateNextWeekCopyWithImpl<UpdateNextWeek>(this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result add(Exercise exercise, OneRm oneRm),
-    @required Result fetch(),
-    @required
-        Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
-    @required Result remove(List<int> ids),
-    @required Result updateNextMonth(int exerciseId, Month nextMonth),
-    @required Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
-  }) {
-    assert(add != null);
-    assert(fetch != null);
-    assert(selectionMode != null);
-    assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
-    return updateNextWeek(exerciseId, nextWeek);
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result add(Exercise exercise, OneRm oneRm),
-    Result fetch(),
-    Result selectionMode(bool isInSelectionMode, List<int> selectedIds),
-    Result remove(List<int> ids),
-    Result updateNextMonth(int exerciseId, Month nextMonth),
-    Result updateNextWeek(int exerciseId, WeekEnum nextWeek),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (updateNextWeek != null) {
-      return updateNextWeek(exerciseId, nextWeek);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result add(Add value),
-    @required Result fetch(Fetch value),
-    @required Result selectionMode(SelectionMode value),
-    @required Result remove(Remove value),
-    @required Result updateNextMonth(UpdateNextMonth value),
-    @required Result updateNextWeek(UpdateNextWeek value),
-  }) {
-    assert(add != null);
-    assert(fetch != null);
-    assert(selectionMode != null);
-    assert(remove != null);
-    assert(updateNextMonth != null);
-    assert(updateNextWeek != null);
-    return updateNextWeek(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result add(Add value),
-    Result fetch(Fetch value),
-    Result selectionMode(SelectionMode value),
-    Result remove(Remove value),
-    Result updateNextMonth(UpdateNextMonth value),
-    Result updateNextWeek(UpdateNextWeek value),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (updateNextWeek != null) {
-      return updateNextWeek(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class UpdateNextWeek implements ExerciseEvent {
-  const factory UpdateNextWeek(
-      {@required int exerciseId,
-      @required WeekEnum nextWeek}) = _$UpdateNextWeek;
-
-  int get exerciseId;
-  WeekEnum get nextWeek;
-  $UpdateNextWeekCopyWith<UpdateNextWeek> get copyWith;
 }
 
 class _$ExerciseStateTearOff {
@@ -1074,26 +666,6 @@ class _$ExerciseStateTearOff {
     );
   }
 
-  MonthUpdateInProgress monthUpdateInProgress() {
-    return const MonthUpdateInProgress();
-  }
-
-  MonthUpdated monthUpdated({@required Month month}) {
-    return MonthUpdated(
-      month: month,
-    );
-  }
-
-  WeekUpdateInProgress weekUpdateInProgress() {
-    return const WeekUpdateInProgress();
-  }
-
-  WeekUpdated weekUpdated({@required Week week}) {
-    return WeekUpdated(
-      week: week,
-    );
-  }
-
   Error error({@required String message}) {
     return Error(
       message: message,
@@ -1116,10 +688,6 @@ mixin _$ExerciseState {
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   });
   @optionalTypeArgs
@@ -1133,10 +701,6 @@ mixin _$ExerciseState {
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   });
@@ -1151,10 +715,6 @@ mixin _$ExerciseState {
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   });
   @optionalTypeArgs
@@ -1168,10 +728,6 @@ mixin _$ExerciseState {
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   });
@@ -1240,10 +796,6 @@ class _$Initial with DiagnosticableTreeMixin implements Initial {
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   }) {
     assert(initial != null);
@@ -1255,10 +807,6 @@ class _$Initial with DiagnosticableTreeMixin implements Initial {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return initial();
   }
@@ -1275,10 +823,6 @@ class _$Initial with DiagnosticableTreeMixin implements Initial {
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   }) {
@@ -1301,10 +845,6 @@ class _$Initial with DiagnosticableTreeMixin implements Initial {
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   }) {
     assert(initial != null);
@@ -1316,10 +856,6 @@ class _$Initial with DiagnosticableTreeMixin implements Initial {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return initial(this);
   }
@@ -1336,10 +872,6 @@ class _$Initial with DiagnosticableTreeMixin implements Initial {
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   }) {
@@ -1406,10 +938,6 @@ class _$AddInProgress with DiagnosticableTreeMixin implements AddInProgress {
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   }) {
     assert(initial != null);
@@ -1421,10 +949,6 @@ class _$AddInProgress with DiagnosticableTreeMixin implements AddInProgress {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return addInProgress();
   }
@@ -1441,10 +965,6 @@ class _$AddInProgress with DiagnosticableTreeMixin implements AddInProgress {
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   }) {
@@ -1467,10 +987,6 @@ class _$AddInProgress with DiagnosticableTreeMixin implements AddInProgress {
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   }) {
     assert(initial != null);
@@ -1482,10 +998,6 @@ class _$AddInProgress with DiagnosticableTreeMixin implements AddInProgress {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return addInProgress(this);
   }
@@ -1502,10 +1014,6 @@ class _$AddInProgress with DiagnosticableTreeMixin implements AddInProgress {
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   }) {
@@ -1569,10 +1077,6 @@ class _$Added with DiagnosticableTreeMixin implements Added {
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   }) {
     assert(initial != null);
@@ -1584,10 +1088,6 @@ class _$Added with DiagnosticableTreeMixin implements Added {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return added();
   }
@@ -1604,10 +1104,6 @@ class _$Added with DiagnosticableTreeMixin implements Added {
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   }) {
@@ -1630,10 +1126,6 @@ class _$Added with DiagnosticableTreeMixin implements Added {
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   }) {
     assert(initial != null);
@@ -1645,10 +1137,6 @@ class _$Added with DiagnosticableTreeMixin implements Added {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return added(this);
   }
@@ -1665,10 +1153,6 @@ class _$Added with DiagnosticableTreeMixin implements Added {
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   }) {
@@ -1738,10 +1222,6 @@ class _$FetchInProgress
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   }) {
     assert(initial != null);
@@ -1753,10 +1233,6 @@ class _$FetchInProgress
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return fetchInProgress();
   }
@@ -1773,10 +1249,6 @@ class _$FetchInProgress
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   }) {
@@ -1799,10 +1271,6 @@ class _$FetchInProgress
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   }) {
     assert(initial != null);
@@ -1814,10 +1282,6 @@ class _$FetchInProgress
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return fetchInProgress(this);
   }
@@ -1834,10 +1298,6 @@ class _$FetchInProgress
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   }) {
@@ -1926,10 +1386,6 @@ class _$Fetched with DiagnosticableTreeMixin implements Fetched {
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   }) {
     assert(initial != null);
@@ -1941,10 +1397,6 @@ class _$Fetched with DiagnosticableTreeMixin implements Fetched {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return fetched(exercises);
   }
@@ -1961,10 +1413,6 @@ class _$Fetched with DiagnosticableTreeMixin implements Fetched {
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   }) {
@@ -1987,10 +1435,6 @@ class _$Fetched with DiagnosticableTreeMixin implements Fetched {
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   }) {
     assert(initial != null);
@@ -2002,10 +1446,6 @@ class _$Fetched with DiagnosticableTreeMixin implements Fetched {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return fetched(this);
   }
@@ -2022,10 +1462,6 @@ class _$Fetched with DiagnosticableTreeMixin implements Fetched {
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   }) {
@@ -2098,10 +1534,6 @@ class _$RemoveInProgress
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   }) {
     assert(initial != null);
@@ -2113,10 +1545,6 @@ class _$RemoveInProgress
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return removeInProgress();
   }
@@ -2133,10 +1561,6 @@ class _$RemoveInProgress
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   }) {
@@ -2159,10 +1583,6 @@ class _$RemoveInProgress
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   }) {
     assert(initial != null);
@@ -2174,10 +1594,6 @@ class _$RemoveInProgress
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return removeInProgress(this);
   }
@@ -2194,10 +1610,6 @@ class _$RemoveInProgress
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   }) {
@@ -2261,10 +1673,6 @@ class _$Removed with DiagnosticableTreeMixin implements Removed {
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   }) {
     assert(initial != null);
@@ -2276,10 +1684,6 @@ class _$Removed with DiagnosticableTreeMixin implements Removed {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return removed();
   }
@@ -2296,10 +1700,6 @@ class _$Removed with DiagnosticableTreeMixin implements Removed {
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   }) {
@@ -2322,10 +1722,6 @@ class _$Removed with DiagnosticableTreeMixin implements Removed {
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   }) {
     assert(initial != null);
@@ -2337,10 +1733,6 @@ class _$Removed with DiagnosticableTreeMixin implements Removed {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return removed(this);
   }
@@ -2357,10 +1749,6 @@ class _$Removed with DiagnosticableTreeMixin implements Removed {
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   }) {
@@ -2450,10 +1838,6 @@ class _$Selected with DiagnosticableTreeMixin implements Selected {
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   }) {
     assert(initial != null);
@@ -2465,10 +1849,6 @@ class _$Selected with DiagnosticableTreeMixin implements Selected {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return selected(selectedIds);
   }
@@ -2485,10 +1865,6 @@ class _$Selected with DiagnosticableTreeMixin implements Selected {
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   }) {
@@ -2511,10 +1887,6 @@ class _$Selected with DiagnosticableTreeMixin implements Selected {
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   }) {
     assert(initial != null);
@@ -2526,10 +1898,6 @@ class _$Selected with DiagnosticableTreeMixin implements Selected {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return selected(this);
   }
@@ -2546,10 +1914,6 @@ class _$Selected with DiagnosticableTreeMixin implements Selected {
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   }) {
@@ -2644,10 +2008,6 @@ class _$Unselected with DiagnosticableTreeMixin implements Unselected {
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   }) {
     assert(initial != null);
@@ -2659,10 +2019,6 @@ class _$Unselected with DiagnosticableTreeMixin implements Unselected {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return unselected(unselectedIds);
   }
@@ -2679,10 +2035,6 @@ class _$Unselected with DiagnosticableTreeMixin implements Unselected {
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   }) {
@@ -2705,10 +2057,6 @@ class _$Unselected with DiagnosticableTreeMixin implements Unselected {
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   }) {
     assert(initial != null);
@@ -2720,10 +2068,6 @@ class _$Unselected with DiagnosticableTreeMixin implements Unselected {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return unselected(this);
   }
@@ -2740,10 +2084,6 @@ class _$Unselected with DiagnosticableTreeMixin implements Unselected {
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   }) {
@@ -2760,726 +2100,6 @@ abstract class Unselected implements ExerciseState {
 
   List<int> get unselectedIds;
   $UnselectedCopyWith<Unselected> get copyWith;
-}
-
-abstract class $MonthUpdateInProgressCopyWith<$Res> {
-  factory $MonthUpdateInProgressCopyWith(MonthUpdateInProgress value,
-          $Res Function(MonthUpdateInProgress) then) =
-      _$MonthUpdateInProgressCopyWithImpl<$Res>;
-}
-
-class _$MonthUpdateInProgressCopyWithImpl<$Res>
-    extends _$ExerciseStateCopyWithImpl<$Res>
-    implements $MonthUpdateInProgressCopyWith<$Res> {
-  _$MonthUpdateInProgressCopyWithImpl(
-      MonthUpdateInProgress _value, $Res Function(MonthUpdateInProgress) _then)
-      : super(_value, (v) => _then(v as MonthUpdateInProgress));
-
-  @override
-  MonthUpdateInProgress get _value => super._value as MonthUpdateInProgress;
-}
-
-class _$MonthUpdateInProgress
-    with DiagnosticableTreeMixin
-    implements MonthUpdateInProgress {
-  const _$MonthUpdateInProgress();
-
-  @override
-  String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
-    return 'ExerciseState.monthUpdateInProgress()';
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties
-      ..add(DiagnosticsProperty('type', 'ExerciseState.monthUpdateInProgress'));
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) || (other is MonthUpdateInProgress);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result initial(),
-    @required Result addInProgress(),
-    @required Result added(),
-    @required Result fetchInProgress(),
-    @required Result fetched(List<Exercise> exercises),
-    @required Result removeInProgress(),
-    @required Result removed(),
-    @required Result selected(List<int> selectedIds),
-    @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
-    @required Result error(String message),
-  }) {
-    assert(initial != null);
-    assert(addInProgress != null);
-    assert(added != null);
-    assert(fetchInProgress != null);
-    assert(fetched != null);
-    assert(removeInProgress != null);
-    assert(removed != null);
-    assert(selected != null);
-    assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
-    assert(error != null);
-    return monthUpdateInProgress();
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result initial(),
-    Result addInProgress(),
-    Result added(),
-    Result fetchInProgress(),
-    Result fetched(List<Exercise> exercises),
-    Result removeInProgress(),
-    Result removed(),
-    Result selected(List<int> selectedIds),
-    Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
-    Result error(String message),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (monthUpdateInProgress != null) {
-      return monthUpdateInProgress();
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result initial(Initial value),
-    @required Result addInProgress(AddInProgress value),
-    @required Result added(Added value),
-    @required Result fetchInProgress(FetchInProgress value),
-    @required Result fetched(Fetched value),
-    @required Result removeInProgress(RemoveInProgress value),
-    @required Result removed(Removed value),
-    @required Result selected(Selected value),
-    @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
-    @required Result error(Error value),
-  }) {
-    assert(initial != null);
-    assert(addInProgress != null);
-    assert(added != null);
-    assert(fetchInProgress != null);
-    assert(fetched != null);
-    assert(removeInProgress != null);
-    assert(removed != null);
-    assert(selected != null);
-    assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
-    assert(error != null);
-    return monthUpdateInProgress(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result initial(Initial value),
-    Result addInProgress(AddInProgress value),
-    Result added(Added value),
-    Result fetchInProgress(FetchInProgress value),
-    Result fetched(Fetched value),
-    Result removeInProgress(RemoveInProgress value),
-    Result removed(Removed value),
-    Result selected(Selected value),
-    Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
-    Result error(Error value),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (monthUpdateInProgress != null) {
-      return monthUpdateInProgress(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class MonthUpdateInProgress implements ExerciseState {
-  const factory MonthUpdateInProgress() = _$MonthUpdateInProgress;
-}
-
-abstract class $MonthUpdatedCopyWith<$Res> {
-  factory $MonthUpdatedCopyWith(
-          MonthUpdated value, $Res Function(MonthUpdated) then) =
-      _$MonthUpdatedCopyWithImpl<$Res>;
-  $Res call({Month month});
-}
-
-class _$MonthUpdatedCopyWithImpl<$Res> extends _$ExerciseStateCopyWithImpl<$Res>
-    implements $MonthUpdatedCopyWith<$Res> {
-  _$MonthUpdatedCopyWithImpl(
-      MonthUpdated _value, $Res Function(MonthUpdated) _then)
-      : super(_value, (v) => _then(v as MonthUpdated));
-
-  @override
-  MonthUpdated get _value => super._value as MonthUpdated;
-
-  @override
-  $Res call({
-    Object month = freezed,
-  }) {
-    return _then(MonthUpdated(
-      month: month == freezed ? _value.month : month as Month,
-    ));
-  }
-}
-
-class _$MonthUpdated with DiagnosticableTreeMixin implements MonthUpdated {
-  const _$MonthUpdated({@required this.month}) : assert(month != null);
-
-  @override
-  final Month month;
-
-  @override
-  String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
-    return 'ExerciseState.monthUpdated(month: $month)';
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties
-      ..add(DiagnosticsProperty('type', 'ExerciseState.monthUpdated'))
-      ..add(DiagnosticsProperty('month', month));
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other is MonthUpdated &&
-            (identical(other.month, month) ||
-                const DeepCollectionEquality().equals(other.month, month)));
-  }
-
-  @override
-  int get hashCode =>
-      runtimeType.hashCode ^ const DeepCollectionEquality().hash(month);
-
-  @override
-  $MonthUpdatedCopyWith<MonthUpdated> get copyWith =>
-      _$MonthUpdatedCopyWithImpl<MonthUpdated>(this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result initial(),
-    @required Result addInProgress(),
-    @required Result added(),
-    @required Result fetchInProgress(),
-    @required Result fetched(List<Exercise> exercises),
-    @required Result removeInProgress(),
-    @required Result removed(),
-    @required Result selected(List<int> selectedIds),
-    @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
-    @required Result error(String message),
-  }) {
-    assert(initial != null);
-    assert(addInProgress != null);
-    assert(added != null);
-    assert(fetchInProgress != null);
-    assert(fetched != null);
-    assert(removeInProgress != null);
-    assert(removed != null);
-    assert(selected != null);
-    assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
-    assert(error != null);
-    return monthUpdated(month);
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result initial(),
-    Result addInProgress(),
-    Result added(),
-    Result fetchInProgress(),
-    Result fetched(List<Exercise> exercises),
-    Result removeInProgress(),
-    Result removed(),
-    Result selected(List<int> selectedIds),
-    Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
-    Result error(String message),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (monthUpdated != null) {
-      return monthUpdated(month);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result initial(Initial value),
-    @required Result addInProgress(AddInProgress value),
-    @required Result added(Added value),
-    @required Result fetchInProgress(FetchInProgress value),
-    @required Result fetched(Fetched value),
-    @required Result removeInProgress(RemoveInProgress value),
-    @required Result removed(Removed value),
-    @required Result selected(Selected value),
-    @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
-    @required Result error(Error value),
-  }) {
-    assert(initial != null);
-    assert(addInProgress != null);
-    assert(added != null);
-    assert(fetchInProgress != null);
-    assert(fetched != null);
-    assert(removeInProgress != null);
-    assert(removed != null);
-    assert(selected != null);
-    assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
-    assert(error != null);
-    return monthUpdated(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result initial(Initial value),
-    Result addInProgress(AddInProgress value),
-    Result added(Added value),
-    Result fetchInProgress(FetchInProgress value),
-    Result fetched(Fetched value),
-    Result removeInProgress(RemoveInProgress value),
-    Result removed(Removed value),
-    Result selected(Selected value),
-    Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
-    Result error(Error value),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (monthUpdated != null) {
-      return monthUpdated(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class MonthUpdated implements ExerciseState {
-  const factory MonthUpdated({@required Month month}) = _$MonthUpdated;
-
-  Month get month;
-  $MonthUpdatedCopyWith<MonthUpdated> get copyWith;
-}
-
-abstract class $WeekUpdateInProgressCopyWith<$Res> {
-  factory $WeekUpdateInProgressCopyWith(WeekUpdateInProgress value,
-          $Res Function(WeekUpdateInProgress) then) =
-      _$WeekUpdateInProgressCopyWithImpl<$Res>;
-}
-
-class _$WeekUpdateInProgressCopyWithImpl<$Res>
-    extends _$ExerciseStateCopyWithImpl<$Res>
-    implements $WeekUpdateInProgressCopyWith<$Res> {
-  _$WeekUpdateInProgressCopyWithImpl(
-      WeekUpdateInProgress _value, $Res Function(WeekUpdateInProgress) _then)
-      : super(_value, (v) => _then(v as WeekUpdateInProgress));
-
-  @override
-  WeekUpdateInProgress get _value => super._value as WeekUpdateInProgress;
-}
-
-class _$WeekUpdateInProgress
-    with DiagnosticableTreeMixin
-    implements WeekUpdateInProgress {
-  const _$WeekUpdateInProgress();
-
-  @override
-  String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
-    return 'ExerciseState.weekUpdateInProgress()';
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties
-      ..add(DiagnosticsProperty('type', 'ExerciseState.weekUpdateInProgress'));
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) || (other is WeekUpdateInProgress);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result initial(),
-    @required Result addInProgress(),
-    @required Result added(),
-    @required Result fetchInProgress(),
-    @required Result fetched(List<Exercise> exercises),
-    @required Result removeInProgress(),
-    @required Result removed(),
-    @required Result selected(List<int> selectedIds),
-    @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
-    @required Result error(String message),
-  }) {
-    assert(initial != null);
-    assert(addInProgress != null);
-    assert(added != null);
-    assert(fetchInProgress != null);
-    assert(fetched != null);
-    assert(removeInProgress != null);
-    assert(removed != null);
-    assert(selected != null);
-    assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
-    assert(error != null);
-    return weekUpdateInProgress();
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result initial(),
-    Result addInProgress(),
-    Result added(),
-    Result fetchInProgress(),
-    Result fetched(List<Exercise> exercises),
-    Result removeInProgress(),
-    Result removed(),
-    Result selected(List<int> selectedIds),
-    Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
-    Result error(String message),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (weekUpdateInProgress != null) {
-      return weekUpdateInProgress();
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result initial(Initial value),
-    @required Result addInProgress(AddInProgress value),
-    @required Result added(Added value),
-    @required Result fetchInProgress(FetchInProgress value),
-    @required Result fetched(Fetched value),
-    @required Result removeInProgress(RemoveInProgress value),
-    @required Result removed(Removed value),
-    @required Result selected(Selected value),
-    @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
-    @required Result error(Error value),
-  }) {
-    assert(initial != null);
-    assert(addInProgress != null);
-    assert(added != null);
-    assert(fetchInProgress != null);
-    assert(fetched != null);
-    assert(removeInProgress != null);
-    assert(removed != null);
-    assert(selected != null);
-    assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
-    assert(error != null);
-    return weekUpdateInProgress(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result initial(Initial value),
-    Result addInProgress(AddInProgress value),
-    Result added(Added value),
-    Result fetchInProgress(FetchInProgress value),
-    Result fetched(Fetched value),
-    Result removeInProgress(RemoveInProgress value),
-    Result removed(Removed value),
-    Result selected(Selected value),
-    Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
-    Result error(Error value),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (weekUpdateInProgress != null) {
-      return weekUpdateInProgress(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class WeekUpdateInProgress implements ExerciseState {
-  const factory WeekUpdateInProgress() = _$WeekUpdateInProgress;
-}
-
-abstract class $WeekUpdatedCopyWith<$Res> {
-  factory $WeekUpdatedCopyWith(
-          WeekUpdated value, $Res Function(WeekUpdated) then) =
-      _$WeekUpdatedCopyWithImpl<$Res>;
-  $Res call({Week week});
-}
-
-class _$WeekUpdatedCopyWithImpl<$Res> extends _$ExerciseStateCopyWithImpl<$Res>
-    implements $WeekUpdatedCopyWith<$Res> {
-  _$WeekUpdatedCopyWithImpl(
-      WeekUpdated _value, $Res Function(WeekUpdated) _then)
-      : super(_value, (v) => _then(v as WeekUpdated));
-
-  @override
-  WeekUpdated get _value => super._value as WeekUpdated;
-
-  @override
-  $Res call({
-    Object week = freezed,
-  }) {
-    return _then(WeekUpdated(
-      week: week == freezed ? _value.week : week as Week,
-    ));
-  }
-}
-
-class _$WeekUpdated with DiagnosticableTreeMixin implements WeekUpdated {
-  const _$WeekUpdated({@required this.week}) : assert(week != null);
-
-  @override
-  final Week week;
-
-  @override
-  String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
-    return 'ExerciseState.weekUpdated(week: $week)';
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties
-      ..add(DiagnosticsProperty('type', 'ExerciseState.weekUpdated'))
-      ..add(DiagnosticsProperty('week', week));
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other is WeekUpdated &&
-            (identical(other.week, week) ||
-                const DeepCollectionEquality().equals(other.week, week)));
-  }
-
-  @override
-  int get hashCode =>
-      runtimeType.hashCode ^ const DeepCollectionEquality().hash(week);
-
-  @override
-  $WeekUpdatedCopyWith<WeekUpdated> get copyWith =>
-      _$WeekUpdatedCopyWithImpl<WeekUpdated>(this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result initial(),
-    @required Result addInProgress(),
-    @required Result added(),
-    @required Result fetchInProgress(),
-    @required Result fetched(List<Exercise> exercises),
-    @required Result removeInProgress(),
-    @required Result removed(),
-    @required Result selected(List<int> selectedIds),
-    @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
-    @required Result error(String message),
-  }) {
-    assert(initial != null);
-    assert(addInProgress != null);
-    assert(added != null);
-    assert(fetchInProgress != null);
-    assert(fetched != null);
-    assert(removeInProgress != null);
-    assert(removed != null);
-    assert(selected != null);
-    assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
-    assert(error != null);
-    return weekUpdated(week);
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result initial(),
-    Result addInProgress(),
-    Result added(),
-    Result fetchInProgress(),
-    Result fetched(List<Exercise> exercises),
-    Result removeInProgress(),
-    Result removed(),
-    Result selected(List<int> selectedIds),
-    Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
-    Result error(String message),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (weekUpdated != null) {
-      return weekUpdated(week);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result initial(Initial value),
-    @required Result addInProgress(AddInProgress value),
-    @required Result added(Added value),
-    @required Result fetchInProgress(FetchInProgress value),
-    @required Result fetched(Fetched value),
-    @required Result removeInProgress(RemoveInProgress value),
-    @required Result removed(Removed value),
-    @required Result selected(Selected value),
-    @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
-    @required Result error(Error value),
-  }) {
-    assert(initial != null);
-    assert(addInProgress != null);
-    assert(added != null);
-    assert(fetchInProgress != null);
-    assert(fetched != null);
-    assert(removeInProgress != null);
-    assert(removed != null);
-    assert(selected != null);
-    assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
-    assert(error != null);
-    return weekUpdated(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result initial(Initial value),
-    Result addInProgress(AddInProgress value),
-    Result added(Added value),
-    Result fetchInProgress(FetchInProgress value),
-    Result fetched(Fetched value),
-    Result removeInProgress(RemoveInProgress value),
-    Result removed(Removed value),
-    Result selected(Selected value),
-    Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
-    Result error(Error value),
-    @required Result orElse(),
-  }) {
-    assert(orElse != null);
-    if (weekUpdated != null) {
-      return weekUpdated(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class WeekUpdated implements ExerciseState {
-  const factory WeekUpdated({@required Week week}) = _$WeekUpdated;
-
-  Week get week;
-  $WeekUpdatedCopyWith<WeekUpdated> get copyWith;
 }
 
 abstract class $ErrorCopyWith<$Res> {
@@ -3553,10 +2173,6 @@ class _$Error with DiagnosticableTreeMixin implements Error {
     @required Result removed(),
     @required Result selected(List<int> selectedIds),
     @required Result unselected(List<int> unselectedIds),
-    @required Result monthUpdateInProgress(),
-    @required Result monthUpdated(Month month),
-    @required Result weekUpdateInProgress(),
-    @required Result weekUpdated(Week week),
     @required Result error(String message),
   }) {
     assert(initial != null);
@@ -3568,10 +2184,6 @@ class _$Error with DiagnosticableTreeMixin implements Error {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return error(message);
   }
@@ -3588,10 +2200,6 @@ class _$Error with DiagnosticableTreeMixin implements Error {
     Result removed(),
     Result selected(List<int> selectedIds),
     Result unselected(List<int> unselectedIds),
-    Result monthUpdateInProgress(),
-    Result monthUpdated(Month month),
-    Result weekUpdateInProgress(),
-    Result weekUpdated(Week week),
     Result error(String message),
     @required Result orElse(),
   }) {
@@ -3614,10 +2222,6 @@ class _$Error with DiagnosticableTreeMixin implements Error {
     @required Result removed(Removed value),
     @required Result selected(Selected value),
     @required Result unselected(Unselected value),
-    @required Result monthUpdateInProgress(MonthUpdateInProgress value),
-    @required Result monthUpdated(MonthUpdated value),
-    @required Result weekUpdateInProgress(WeekUpdateInProgress value),
-    @required Result weekUpdated(WeekUpdated value),
     @required Result error(Error value),
   }) {
     assert(initial != null);
@@ -3629,10 +2233,6 @@ class _$Error with DiagnosticableTreeMixin implements Error {
     assert(removed != null);
     assert(selected != null);
     assert(unselected != null);
-    assert(monthUpdateInProgress != null);
-    assert(monthUpdated != null);
-    assert(weekUpdateInProgress != null);
-    assert(weekUpdated != null);
     assert(error != null);
     return error(this);
   }
@@ -3649,10 +2249,6 @@ class _$Error with DiagnosticableTreeMixin implements Error {
     Result removed(Removed value),
     Result selected(Selected value),
     Result unselected(Unselected value),
-    Result monthUpdateInProgress(MonthUpdateInProgress value),
-    Result monthUpdated(MonthUpdated value),
-    Result weekUpdateInProgress(WeekUpdateInProgress value),
-    Result weekUpdated(WeekUpdated value),
     Result error(Error value),
     @required Result orElse(),
   }) {
