+++ lib/application/exercise/month/month_bloc.dart
@@ -1,0 +1,53 @@
+import 'dart:async';
+
+import 'package:bloc/bloc.dart';
+import 'package:dartz/dartz.dart';
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:meta/meta.dart';
+
+import 'package:power_progress/domain/core/entities/value_objects/month.dart';
+import 'package:power_progress/domain/exercise/entities/exercise_failure.dart';
+import 'package:power_progress/domain/exercise/usecases/update_exercise_next_month.dart';
+
+part 'month_event.dart';
+part 'month_state.dart';
+part 'month_bloc.freezed.dart';
+
+class MonthBloc extends Bloc<MonthEvent, MonthState> {
+  final UpdateExerciseNextMonth updateExerciseNextMonth;
+
+  MonthBloc({
+    @required this.updateExerciseNextMonth,
+  });
+
+  @override
+  MonthState get initialState => const MonthState.initial();
+
+  @override
+  Stream<MonthState> mapEventToState(
+    MonthEvent event,
+  ) async* {
+    yield* _handleUpdateNextMonthEvent(event);
+  }
+
+  Stream<MonthState> _handleUpdateNextMonthEvent(MonthEvent event) async* {
+    yield const MonthState.monthUpdateInProgress();
+
+    final output = await updateExerciseNextMonth(
+      UpdateExerciseNextMonthParams(
+        exerciseId: event.exerciseId,
+        nextMonth: event.nextMonth,
+      ),
+    );
+
+    Stream<MonthState> onFailure(ExerciseFailure failure) async* {
+      yield MonthState.error(message: failure.toErrorMessage());
+    }
+
+    Stream<MonthState> onSuccess(Unit unit) async* {
+      yield MonthState.monthUpdated(month: event.nextMonth);
+    }
+
+    yield* output.fold(onFailure, onSuccess);
+  }
+}
