+++ lib/domain/exercise/entities/exercise_failure.dart
@@ -1,5 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:flutter/foundation.dart';
+import 'package:power_progress/core/messages/errors.dart';
 
 part 'exercise_failure.freezed.dart';
 
@@ -9,4 +10,16 @@ abstract class ExerciseFailure with _$ExerciseFailure {
   const factory ExerciseFailure.unexpectedError() = UnexpectedError;
   const factory ExerciseFailure.oneRmDoestNotExist() = OneRmDoestNotExist;
   const factory ExerciseFailure.oneRmAlreadyExists() = OneRmAlreadyExists;
+}
+
+/// Extension methods of the exercise failure class
+extension WeeksX on ExerciseFailure {
+  String toErrorMessage() {
+    return when(
+      storageError: () => storageErrorMessage,
+      unexpectedError: null,
+      oneRmDoestNotExist: null,
+      oneRmAlreadyExists: null,
+    );
+  }
 }
